// src/app/private-app/interfaces/compromiso-recepcion.ts
export interface CompromisoRecepcion {
  razon_social: string;
  representante_legal: string;
  area_dedicacion: string;
  telefono: string;
  direccion: string;
  dias_habiles: string;
  horario: string;
  nombre_representante: string;
  funcion: string;
  telefono_representante: string;
  email_representante: string;
  nombre_estudiante: string;
  area_estudiante: string;
}